Move Preview option test to ConfigureFreestyleProject spec
